Updated to support multiple format strings
Users can now pass a variadic list of format strings to the NewVersion function.
I also added a new example file for Series function.

I also updated some funcs to remove some unnecessary error returns. For example
the regex used to extract the version parts ensure the major, minor, and micro
are numbers. This guarantee made the func simpler and remove the need for an
error to be returned.

This is also simplified the Compare func as it now no longer returns an error.
NewCollection also accepts multiple format strings since it was essentially just
a wrapper around NewVersion.

Finally note that I changed the order of version and format parameters in the
NewVersion func. It just seemed unnatural. When I was using the code downstream
in some other work, I noticed that I would subconsciously pass the version and
then the format so I propagated that here.

closes: https://github.com/shazib-summar/go-calver/issues/16
closes: https://github.com/shazib-summar/go-calver/issues/17

Signed-off-by: Shazib <shazib.summar.skipq@gmail.com>