Added ParseWithOptions and NewCollectionWithOptions
These funcs allow taking in a list of options that define the behavior of the
function. Today there's only one option `WithFormat` that takes a list of
formats.

For example, if we have the following versions:

```
2025-07-14
2025-07-15
```

We can obtain a Version object with the following code:

```go
ver, err := calver.ParseWithOptions(
    "2025-07-14",
    calver.WithFormat(
        "<YYYY>-<MM>-<DD>",
        "<YYYY>.<MM>.<DD>",
    ),
)
```

We can create a `Collection` with the following code:

```go
collection, err := calver.NewCollectionWithOptions(
    []string{"2025-07-14", "2025-07-15"},
    calver.WithFormat(
        "<YYYY>-<MM>-<DD>",
        "<YYYY>.<MM>.<DD>",
    ),
)
```

Also added more tests and updated documentation. Added an example for the
`Series` method.

Signed-off-by: Shazib <shazib.summar.skipq@gmail.com>